requirements.txt
beautifulsoup4
fastapi
httpx
jinja2
python-dotenv
supabase
uvicorn
